pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blood-foundation"
version = "0.1.0"
description = "Modular transformer-based foundation model skeleton for blood-test tabular data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "pandas>=2.0",
  "numpy>=1.24",
  "torch>=2.1",
  "matplotlib>=3.7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
